--- templates/apply_pattern.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<inkscape-extension xmlns="http://www.inkscape.org/namespace/inkscape/extension">
-    <name>{% trans %}Apply Pattern{% endtrans %}</name>
-    <id>org.inkstitch.apply_pattern.{{ locale }}</id>
-    <param name="extension" type="string" gui-hidden="true">apply_pattern</param>
-    <effect>
-        <object-type>all</object-type>
-        <effects-menu>
-            <submenu name="Ink/Stitch" />
-        </effects-menu>
-    </effect>
-    <script>
-        {{ command_tag | safe }}
-    </script>
-</inkscape-extension>
