+++ templates/selection_to_pattern.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<inkscape-extension xmlns="http://www.inkscape.org/namespace/inkscape/extension">
+    <name>{% trans %}Selection to pattern{% endtrans %}</name>
+    <id>org.inkstitch.selection_to_pattern.{{ locale }}</id>
+    <param name="extension" type="string" gui-hidden="true">selection_to_pattern</param>
+    <effect>
+        <object-type>all</object-type>
+        <effects-menu>
+            <submenu name="Ink/Stitch">
+                <submenu name="{% trans %}Edit{% endtrans %}" />
+            </submenu>
+        </effects-menu>
+    </effect>
+    <script>
+        {{ command_tag | safe }}
+    </script>
+</inkscape-extension>
